pyproject.toml
[tool.poetry]
name = "vacancies-review"
version = "0.1.0"
description = ""
authors = ["aleksandr.rybin <aleksandr.rybin@glowbyteconsulting.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
pydantic-settings = "^2.3.2"
asyncpg = "^0.29.0"
black = "^24.4.2"
ruff = "^0.4.10"
fastapi = "^0.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
